added feature for create and edit user with role!

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 Use App\Http\Requests\UserRequest as Request;
 use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Models\Role;
 use App\User;
 
 class UsersController extends Controller
@@ -31,7 +32,7 @@ class UsersController extends Controller
     public function create()
     {
         if (Auth::user()->hasPermissionTo('write users')) {
-            return view('user.create');
+            return view('user.create')->with('roles', Role::all());
         } else {
             flash(trans("permission.write.user"))->error()->important();
             return back();
@@ -46,7 +47,9 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
-        User::create($request->all());
+        $user = User::create($request->all());
+        $user->assignRole(Role::find($request->role_id)->name);
+        flash(trans("messages.success.user.store"))->success()->important();
         return redirect()->route('user.index');
     }
 
@@ -75,7 +78,7 @@ class UsersController extends Controller
     public function edit($id)
     {
         if (Auth::user()->hasPermissionTo('write users')) {
-            return view('user.edit')->with('user', User::find($id));
+            return view('user.edit')->with('user', User::find($id))->with('roles', Role::all());
         } else {
             flash(trans("permission.write.user"))->error()->important();
             return back();
@@ -91,7 +94,16 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
-        User::find($id)->fill($request->all())->save();
+        $user = User::find($id)->fill($request->all());
+        if($request->password) {
+            $user->password = bcrypt($request->password);
+        }
+        $user->save();
+        if (User::role('admin')->get()->count() == 1 && User::role('admin')->get()->first()->id == $user->id)
+            flash(trans("messages.last_admin"))->error()->important();
+        else
+            $user->syncRoles([Role::find($request->role_id)->name]);
+        flash(trans("messages.success.user.update"))->success()->important();
         return redirect()->route('user.index');
     }
 
@@ -105,13 +117,21 @@ class UsersController extends Controller
     {
         if (Auth::user()->hasPermissionTo('delete users')) {
             $user = User::withTrashed()->find($id);
-            if($user->trashed())
-                $user->restore();
-            else
-                $user->delete();
-            return redirect()->back();
+
+            if (User::role('admin')->get()->count() == 1 && User::role('admin')->get()->first()->id == $user->id)
+                flash(trans("messages.last_admin"))->error()->important();
+            else {
+                if($user->trashed()) {
+                    $user->restore();
+                    flash(trans("messages.success.user.restore"))->info()->important();
+                } else {
+                    $user->delete();
+                    flash(trans("messages.success.user.delete"))->warning()->important();
+                }
+                return redirect()->back();
+            }
         } else {
-            flash(trans("permission.delete.users"))->error()->important();
+            flash(trans("permission.delete.user"))->error()->important();
             return back();
         }
     }
